src/main/java/com/my/spring5recipeapp/converters/IngredientToIngredientCommand.java
package com.my.spring5recipeapp.converters;

import com.my.spring5recipeapp.commands.IngredientCommand;
import com.my.spring5recipeapp.domain.Ingredient;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class IngredientToIngredientCommand implements Converter<Ingredient, IngredientCommand> {
   private final UnitOfMeasureToUnitOfMeasureCommand uomConverter;

    public IngredientToIngredientCommand(UnitOfMeasureToUnitOfMeasureCommand uomConverter) {
        this.uomConverter = uomConverter;
    }

    @Synchronized
    @Nullable
    @Override
    public IngredientCommand convert(Ingredient ingredient) {
        if (ingredient==null) {
            return null;
        }
        final IngredientCommand ingredientCommand=new IngredientCommand();
        ingredientCommand.setId(ingredient.getId());
        ingredientCommand.setDescription(ingredient.getDescription());
        if(ingredient.getRecipe()!=null) {
            ingredientCommand.setRecipeId(ingredient.getRecipe().getId());
        }
        ingredientCommand.setUom(uomConverter.convert(ingredient.getUom()));
        ingredientCommand.setAmount(ingredient.getAmount());
        return ingredientCommand;
    }
}
